requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
autograd==1.5
autoray==0.6.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
cirq-core==1.3.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
deprecation==2.1.0
dill==0.4.0
duet==0.2.9
et_xmlfile==2.0.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
future==1.0.0
gast==0.4.0
google-auth==2.39.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.71.0
h5py==3.13.0
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
keras==2.11.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.4.3
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
openfermion==1.6.1
openpyxl==3.1.5
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pbr==6.1.1
PennyLane==0.31.1
PennyLane-Lightning==0.32.0
pillow==11.2.1
prettytable==3.16.0
propcache==0.3.1
protobuf==3.19.6
psutil==7.0.0
PubChemPy==1.0.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.0.9
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qiskit==1.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
rustworkx==0.16.0
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
six==1.17.0
some-package==0.1
sortedcontainers==2.4.0
stevedore==5.4.1
symengine==0.14.0
sympy==1.13.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorcircuit==0.10.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
tensornetwork==0.4.6
termcolor==3.0.1
toml==0.10.2
torch==2.0.1
torch-geometric==2.6.1
tqdm==4.66.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.19.0
zipp==3.21.0
